파일구조 변경 및 버그수정
