+++ src/pages/passwordReset/ResetEmailCheck.tsx
@@ -1,31 +1,20 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate, useLocation } from "react-router-dom";
-import { useMutation } from "react-query";
-import styles from "../style/emailVerify.module.css";
-import deplogLogo from "../images/deplogLogo.png";
-import { constants } from "../constants";
-import { RegisterReq } from '../api/RegisterReq';
-import { sendMail } from '../api/RegisterReq';
+import { useLocation, useNavigate } from "react-router-dom";
+import styles from "../../style/email/emailVerify.module.css";
+import deplogLogo from "../../images/deplogLogo.png";
+import { constants } from "../../constants/logIn";
+import { sendMail, checkMail } from '../../api/PassReset';
 
-const RegisterEmailVerify: React.FC = () => {
+const ResetEmailCheck: React.FC = () => {
+  const location = useLocation();
+  const { email } = location.state as { email: string }; 
   const [verificationMessage, setVerificationMessage] = useState("");
   const [resendEnabled, setResendEnabled] = useState(true);
   const [time, setTime] = useState<number>(180);
   const navigate = useNavigate();
-  const location = useLocation();
-  const { email, password, name, part, generation } = location.state as { email: string; password: string; name: string; part: string; generation: number }; // Get all necessary data from state
-
-  const { mutate: registerUser } = useMutation(RegisterReq, {
-    onSuccess: () => {
-      navigate('/emailSuccess'); 
-    },
-    onError: (error: any) => {
-      setVerificationMessage(constants.verifyCompleteMessage);
-      //setVerificationMessage('회원가입 중 오류가 발생했습니다: ' + (error.response?.data?.message || error.message)); 
-    }
-  });
 
   useEffect(() => {
+    //JS 타이머 로직
     let interval: number | null = null;
     if (!resendEnabled) {
       interval = window.setInterval(() => {
@@ -43,19 +32,27 @@ const RegisterEmailVerify: React.FC = () => {
 
   const handleVerify = async () => {
     try {
-      registerUser({ email, password, name, part, generation });
+      const response = await checkMail(email);
+      if (response.data.verified) {
+        alert(constants.emailSuccessMessage);
+        navigate('/changingPassword', { state: { email } }); // 이메일을 state로 전달하고 비밀번호 변경 페이지로 이동
+      } else {
+        setVerificationMessage(constants.verifyCompleteMessage);
+      }
     } catch (error) {
-      alert(constants.verifyErrorMessage);
+      console.error("Failed to verify email", error);
+      alert(constants.emailFailMessage);
     }
   };
+  
 
   const handleResend = async () => {
     setResendEnabled(false);
     setTime(180);
     try {
-      await sendMail(email);
+      await sendMail(email); // 전달받은 이메일로 재전송
     } catch (error) {
-      alert(constants.resendMailErrorMessage);
+      console.error("Failed to resend email", error);
     }
   };
 
@@ -81,4 +78,4 @@ const RegisterEmailVerify: React.FC = () => {
   );
 };
 
-export default RegisterEmailVerify;
+export default ResetEmailCheck;
