+++ src/style/email/emailVerify.module.css
@@ -1,0 +1,143 @@
+@import url("https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap");
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+html,
+body {
+  height: 100%;
+  font-family: "Noto Sans KR", sans-serif;
+}
+
+.centeredContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  width: 100vw;
+  background-color: #f5f5f5;
+}
+
+.verifyForm {
+  width: 508px;
+  padding: 35px 23px;
+  background-color: #e9ecef;
+  border-radius: 8px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+}
+
+.logoContainer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 57.42px;
+  padding: 8px 4px;
+  gap: 4px;
+}
+
+.logo {
+  width: 143.58px;
+  height: 41.42px;
+}
+
+.titleContainer {
+  display: flex;
+  box-sizing: border-box;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 50px;
+  padding: 16px 4px 4px 4px;
+}
+
+.title {
+  font-size: 21.33px;
+  line-height: 29.86px;
+  font-weight: 700;
+  color: #000000;
+}
+
+.subtitleContainer {
+  display: flex;
+  box-sizing: border-box;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 46px;
+  padding: 4px;
+  text-align: center;
+}
+
+.subtitle {
+  font-size: 14px;
+  color: #000000;
+  font-weight: 400;
+  line-height: 19.2px;
+}
+
+.errorMessageContainer {
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 27px;
+  padding: 4px;
+  gap: 4px;
+}
+
+.errorMessage {
+  color: red;
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 19.2px;
+}
+
+.buttonGroup {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+  width: 100%;
+}
+
+.verifyButton {
+  width: 100%;
+  height: 58px;
+  padding: 16px 12px;
+  background-color: #050c20;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 16px;
+  font-family: "Noto Sans KR", sans-serif;
+}
+
+.resendButton {
+  width: 100%;
+  height: 58px;
+  padding: 16px 12px;
+  background-color: #050c20;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 16px;
+  font-family: "Noto Sans KR", sans-serif;
+}
+
+.resendButtonDisabled {
+  width: 100%;
+  height: 58px;
+  padding: 16px 12px;
+  background-color: #858585;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: not-allowed;
+  font-size: 16px;
+}
