+++ src/style/logIn/LoginForm.module.css
@@ -1,0 +1,205 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap');
+
+.centeredContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width: 100vw;
+    background-color: #f5f5f5;
+}
+
+.loginForm {
+    width: 509px;
+    height: 688px;
+    padding: 45px 23px;
+    background-color: #e9ecef; 
+}
+
+.logoContainer {
+    box-sizing: border-box;
+    height: 72.72px;
+    margin-top: 77.64px;
+    padding: 8px 4px;
+}
+
+.logo {
+    display: block;
+    margin: 0 auto;
+    height: 56.72px;
+}
+
+.labelContainer {
+    width: 100%;
+    height: 46px;
+    padding: 12px 4px;
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    box-sizing: border-box;
+}
+
+.labelContainer label {
+    font-size: 16px;
+    line-height: 22.4px;
+    font-weight: 700;
+    color: #000000;
+    margin: 0;
+    padding: 0;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.formGroup input, .errorMessageContainer {
+    width: 100%;
+    padding: 16px 12px;
+    border: solid 1px rgba(173, 187, 200, 0.25);
+    border-radius: 2px;
+    box-sizing: border-box;
+    background-color: white;
+}
+
+.formGroup input {
+    color:#6B6B6B;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 25.6px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.error {
+    border-color: red;
+}
+
+.errorMessageContainer {
+    height: 27px;
+    min-height: 27px; /* 공백 크기 유지를 위한 최소 높이 설정 */
+    padding: 4px;
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    border: 0;
+    background-color: #e9ecef; /* Optional: to highlight the error message container */
+    box-sizing: border-box;
+}
+
+.errorMessage {
+    color: red;
+    font-size: 12px;
+}
+
+.capslockWarningContainer {
+    padding: 4px 4px 12px 4px;
+    box-sizing: border-box;
+    height: 39px;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+
+.capslockWarningTextContainer {
+    padding: 0px 4px 4px 4px;
+    gap: 4px;
+    height: 27px;
+}
+
+.capslockWarning {
+    color: #FF1C00;
+    font-size: 12px;
+    font-weight: 700;
+    line-height: 19.2px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.passwordContainer {
+    position: relative;
+    display: flex;
+    align-items: center;
+}
+
+.passwordToggleIcon {
+    position: absolute;
+    top: 17px;
+    right: 12px;
+    cursor: pointer;
+    color: #aaa;
+}
+
+.passwordToggleIcon img {
+    width: 24px;
+    height: 24px;
+}
+
+.loginButton {
+    width: 100%;
+    height: 58px;
+    padding: 16px 12px;
+    background-color: #050C20; /* Gray button color */
+    color: white;
+    border: 1px;
+    border-radius: 2px;
+    cursor: pointer;
+    gap: 4px;
+    box-sizing: border-box;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 22.4px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.loginButton:disabled {
+    background-color: #858585;
+    cursor: not-allowed;
+}
+
+.autoLoginContainer {
+    padding: 12px 4px 12px 4px;
+    box-sizing: border-box;
+    height: 51px;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+
+.autoLoginContainer.noPaddingBottom {
+    height:39px;
+    padding-bottom: 0px;
+}
+
+.checkbox {
+    width: 28px;
+    height: 28px; 
+    cursor: pointer;
+}
+
+.autoLoginLabel {
+    font-size: 12px;
+    font-weight: 400;
+    color: #6B6B6B;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.links {
+    display: flex;
+    box-sizing: border-box;
+    margin-top: 25px;
+    margin-bottom: 81.64px;
+    width: 100%;
+    height: 56px;
+    padding: 16px 12px;
+    gap: 4px;
+    justify-content: space-between;
+}
+
+.links span {
+    color: #6B6B6B;
+    text-decoration: none;
+    cursor: pointer;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 24px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.links a:hover, .links span:hover {
+    text-decoration: underline;
+}
