+++ src/style/register/RegisterPage.module.css
@@ -1,0 +1,250 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap');
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+html, body {
+    height: 100%;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.centeredContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width: 100vw;
+    background-color: #f5f5f5;
+}
+
+.registerForm {
+    width: 508px;
+    padding: 35px 23px;
+    background-color: #e9ecef;
+}
+
+.logoContainer {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 57.42px;
+    padding: 8px 4px;
+    gap: 4px;
+}
+
+.logo {
+    width: 143.58px;
+    height: 41.42px;
+}
+
+.titleContainer {
+    display: flex;
+    box-sizing: border-box;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 50px;
+    padding: 16px 4px 4px 4px;
+}
+
+.title {
+    font-size: 21.33px;
+    line-height: 29.86px;
+    font-weight: 700;
+    color: #000000;
+}
+
+.subtitleContainer {
+    display: flex;
+    box-sizing: border-box;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 27px;
+    padding: 4px;
+}
+
+.subtitle {
+    font-size: 12px;
+    color: #000000;
+    font-weight: 400;
+    line-height: 19.2px;
+}
+
+.labelContainer {
+    width: 100%;
+    padding: 12px 4px;
+    box-sizing: border-box;
+    height: 46px;
+    display: flex;
+    align-items: center;
+    gap: 4px;
+}
+
+.labelContainer label {
+    font-size: 16px;
+    line-height: 22.4px;
+    font-weight: 700;
+    color: #000000;
+}
+
+.inputWithButton {
+    display: flex;
+    gap: 2px;
+}
+
+.inputWithButton input {
+    box-sizing: border-box;
+    width: 373px;
+    height: 58px;
+    flex: 1;
+    padding: 16px 12px;
+    border: solid 1px #ADBBC8;
+    border-radius: 2px;
+    background-color: white;
+}
+
+input {
+    color: #000000;
+}
+
+.error {
+    border-color: red;
+}
+
+.errorMessageContainer {
+    height: 27px;
+    padding: 4px;
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    background-color: #e9ecef;
+}
+
+.errorMessage {
+    color: red;
+    font-size: 12px;
+    font-family: 'Noto Sans KR', sans-serif;
+    line-height: 19.2px;
+    font-weight: 400;
+}
+
+.passwordContainer {
+    position: relative;
+    display: flex;
+    align-items: center;
+}
+
+.passwordContainer input {
+    box-sizing: border-box;
+    width: 373px;
+    height: 58px;
+    flex: 1;
+    padding: 16px 12px;
+    border: solid 1px #ADBBC8;
+    border-radius: 2px;
+    background-color: white;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 25.6px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.passwordToggleIcon {
+    position: absolute;
+    top: 17px;
+    right: 12px;
+    cursor: pointer;
+    color: #aaa;
+}
+
+.passwordToggleIcon img {
+    width: 24px;
+    height: 24px;
+}
+
+.checkButton {
+    padding: 16px 12px;
+    box-sizing: border-box;
+    width: 88.5px;
+    height: 58px;
+    background-color: #858585;
+    color: #FFFFFF;
+    border: none;
+    cursor: not-allowed;
+    border-radius: 2px;
+    gap: 3px;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 22.4px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.checkButton.enabled {
+    background-color: #050C20;
+    cursor: pointer;
+}
+
+.buttonGroup {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 13px;
+}
+
+.cancelButton {
+    width: 154px;
+    height: 58px;
+    padding: 16px 12px;
+    background-color: #e9ecef;
+    color: #858585;
+    border: none;
+    border-radius: 2px;
+    cursor: pointer;
+    font-size: 16px;
+    font-weight: 500;
+    line-height: 24px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.nextButton {
+    width: 306px;
+    height: 58px;
+    padding: 16px 12px;
+    background-color: #050C20;
+    color: white;
+    border: none;
+    border-radius: 2px;
+    cursor: pointer;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 22.4px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.inputWithButton input {
+    box-sizing: border-box;
+    width: 373px;
+    height: 58px;
+    flex: 1;
+    padding: 16px 12px;
+    border: solid 1px #ADBBC8;
+    border-radius: 2px;
+    background-color: white;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 25.6px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.cancelButton:hover {
+    color: #050C20;
+}
+
+.nextButton:disabled {
+    background-color: #858585;
+    cursor: not-allowed;
+}
