+++ src/style/search/searchResultsPage.module.css
